requirements.txt
discord.py[voice]
peony-twitter
psutil
pyyaml
TwitchAPI
youtube_dl
PyNaCl
youtube_dl